requirements-colab.txt
accelerate
cmudict
cn2an
g2p_en
gradio>=4.32
jieba
librosa==0.9.2
loguru
nltk<=3.8.1
num2words
numpy<2
onnx
onnxconverter-common
onnxruntime
onnxruntime-gpu
onnxsim-prebuilt
pyannote.audio>=3.1.0
pyloudnorm
pyopenjtalk-dict
pypinyin
pyworld-prebuilt
torch
torchaudio
torchvision
transformers>=4.30,<4.46
umap-learn
